pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "speech-decoding"
version = "0.1.0"
description = "Cross-patient speech decoding from intra-operative uECOG"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0",
    "mne>=1.6",
    "mne-bids>=0.14",
    "numpy>=1.24",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "pyyaml>=6.0",
    "h5py>=3.9",
    "tqdm>=4.65",
    "transformers>=4.40",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests that load real BIDS data (deselect with '-m \"not slow\"')",
]
